RuneRPG - Script-system updated - Renaming some variables - Fixes to GlobalUI system, Escape query (all queries) now are laid on top of the global ui. - Pause-menu added to cutscenes - Cutscene game state added to handle the in-cutscene menus. - Removed old projects that are deprecated and not used any more. - New System/DataTypes.h header added to define better names for certain types (e.g. ing64 for 'long long' or '__int64')

+++ src/Graphics/GraphicsManager.h
@@ -15,6 +15,7 @@
 #include "Graphics/Messages/GraphicsMessage.h"
 #include "ShaderManager.h"
 #include "VFCOctree.h"
+#include "System/DataTypes.h"
 
 struct RenderSettings;
 class Texture;
@@ -208,7 +209,7 @@ public:
 	const float PhysicsFrameTime() const { return physicsFrameTime; };
     float renderViewportsFrameTime;
     float renderViewportFrameTime[4];
-    float graphicsMessageProcessingFrameTime,
+    int64 graphicsMessageProcessingFrameTime,
         graphicsUpdatingFrameTime,
         preRenderFrameTime,
         swapBufferFrameTime,
