+++ src/Graphics/Render/RenderViewport.h
@@ -18,7 +18,7 @@ class RenderViewport : public Viewport {
 public:
 	/// Default constructor, pass optional gui source to attach a gui to it on creation.
 	RenderViewport(String uiSource);
-	RenderViewport(float leftX, float bottomY, float width, float height);
+	RenderViewport(int leftX, int bottomY, int width, int height);
 	virtual ~RenderViewport();
 
 	/// Sets the viewport to use relative coordinates or absolute.
