+++ src/Projects/RuneRPG/GameStates/GlobalState/RuneGlobalState.h
@@ -27,7 +27,7 @@ public:
 	/// Input functions for the various states
 	void MouseClick(bool down, int x = -1, int y = -1, UIElement * elementClicked = NULL);
 	void MouseRightClick(bool down, int x = -1, int y = -1, UIElement * elementClicked = NULL);
-	virtual void MouseMove(float x, float y, bool lDown = false, bool rDown = false, UIElement * elementOver = NULL);
+	virtual void MouseMove(int x, int y, bool lDown = false, bool rDown = false, UIElement * elementOver = NULL);
 	void MouseWheel(float delta);
 	/// Interpret selection queries for additional processing (updating UI for example).
 	void OnSelect(Selection &selection);
