+++ src/Projects/RuneRPG/GameStates/Map/MapState.h
@@ -40,7 +40,7 @@ public:
 	/// Input functions for the various states
 	void MouseClick(bool down, int x = -1, int y = -1, UIElement * elementClicked = NULL);
 	void MouseRightClick(bool down, int x = -1, int y = -1, UIElement * elementClicked = NULL);
-	virtual void MouseMove(float x, float y, bool lDown = false, bool rDown = false, UIElement * elementOver = NULL);
+	virtual void MouseMove(int x, int y, bool lDown = false, bool rDown = false, UIElement * elementOver = NULL);
 	void MouseWheel(float delta);
 	/// Callback from the Input-manager, query it for additional information as needed.
 	void KeyPressed(int keyCode, bool downBefore);
